Ya funciona los gets con estudiantes sin mascota pero no he conseguido que a un estudiante sin mascota con el metodo put se pueda añadir una mascota

+++ src/main/java/com/benjamin/hogwarts/service/impl/EstudianteServiceImpl.java
@@ -66,7 +66,9 @@ public class EstudianteServiceImpl implements EstudianteService {
         Mascota mascota = estudianteExistente.getMascota();
         estudianteMapper.updateEntityFromDto(dto, estudianteExistente);
         Estudiante estudianteActualizado = estudianteRespository.save(estudianteExistente);
-        mascotaRepository.delete(mascota);
+        if (dto.getMascota() == null ){
+            mascotaRepository.delete(mascota);
+        }
         return estudianteMapper.toDTO(estudianteActualizado);
     }
     @Override
